Added solver related tests to test_solvers.py from test_matrices.py
